React Asssignment updated

+++ ReST_Assignment/Entity/Employee.cs
@@ -7,7 +7,7 @@
 
         public string City { get; set; }
 
-        public int Salary { get; set; }
+        public double Salary { get; set; }
         
     }
 }
